% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangles.R
\name{aggregate_loss_data}
\alias{aggregate_loss_data}
\title{Aggregate Loss data}
\usage{
aggregate_loss_data(claim_dat, limit = NA)
}
\arguments{
\item{claim_dat}{claims data}

\item{limit}{optional limit}
}
\value{
df
}
\description{
Aggregate Loss data
}
