% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{auto_claims}
\alias{auto_claims}
\title{auto_claims}
\format{
A \code{data.frame} with 404704 rows and 18 variables:
\describe{
\item{\code{accident_year}}{integer. DESCRIPTION.}
\item{\code{claim_id}}{integer. DESCRIPTION.}
\item{\code{injury_level}}{integer. DESCRIPTION.}
\item{\code{accident_date}}{double. DESCRIPTION.}
\item{\code{accident_hour}}{integer. DESCRIPTION.}
\item{\code{accident_in_city}}{integer. DESCRIPTION.}
\item{\code{accident_road_type}}{integer. DESCRIPTION.}
\item{\code{accident_location_id}}{integer. DESCRIPTION.}
\item{\code{accident_road_section_id}}{integer. DESCRIPTION.}
\item{\code{accident_location_type}}{integer. DESCRIPTION.}
\item{\code{accident_cause}}{integer. DESCRIPTION.}
\item{\code{accident_event_type}}{integer. DESCRIPTION.}
\item{\code{accident_weather}}{integer. DESCRIPTION.}
\item{\code{accident_traffic}}{integer. DESCRIPTION.}
\item{\code{accident_surface_conditions}}{integer. DESCRIPTION.}
\item{\code{accident_surface}}{integer. DESCRIPTION.}
\item{\code{accident_lat}}{double. DESCRIPTION.}
\item{\code{accident_lon}}{double. DESCRIPTION.}
}
}
\usage{
auto_claims
}
\description{
Auto Liability Claims Data from original source \link[trafficaccidents:event]{trafficaccidents::event}
}
\keyword{datasets}
