% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-data.R
\docType{data}
\name{claims_transactional}
\alias{claims_transactional}
\title{claims_transactional}
\format{
A \code{data.frame} with 80278 rows and 12 variables:
\describe{
\item{\code{claim_num}}{integer. DESCRIPTION.}
\item{\code{claim_id}}{character. DESCRIPTION.}
\item{\code{accident_date}}{double. DESCRIPTION.}
\item{\code{state}}{character. DESCRIPTION.}
\item{\code{claimant}}{character. DESCRIPTION.}
\item{\code{report_date}}{double. DESCRIPTION.}
\item{\code{status}}{character. DESCRIPTION.}
\item{\code{payment}}{double. DESCRIPTION.}
\item{\code{case}}{double. DESCRIPTION.}
\item{\code{transaction_date}}{double. DESCRIPTION.}
\item{\code{trans_num}}{integer. DESCRIPTION.}
\item{\code{paid}}{double. DESCRIPTION.}
}
}
\usage{
claims_transactional
}
\description{
Transactional claims dataset. Convert to a static \code{lossrun} using the
\code{\link[=loss_run]{loss_run()}} function.
}
\keyword{datasets}
