% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-connect.R
\name{connect_db}
\alias{connect_db}
\title{Connect DB}
\usage{
connect_db(pool = TRUE)
}
\arguments{
\item{pool}{Logical - should the connection be pooled?}
}
\value{
a database connection
}
\description{
Connect to the Actuarial Database Instance.
}
