% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-create_tbl.R
\name{create_tbl}
\alias{create_tbl}
\title{Create Table}
\usage{
create_tbl(
  conn,
  tbl_name,
  csv_path = "data-raw/database/CSV",
  sql_path = "data-raw/database/SQL",
  drop_if_exists = TRUE
)
}
\arguments{
\item{conn}{database connection}

\item{tbl_name}{character string representing table name}

\item{csv_path}{base path (excluding file) to the CSV file}

\item{sql_path}{base path (excluding file) to the SQL file}

\item{drop_if_exists}{Should the table be dropped (with CASCADE) if it already exists?}
}
\value{
the created database table returned as an R \code{\link[=data.frame]{data.frame()}}.
}
\description{
Creates a database table given a connection, table name, and:
\itemize{
\item SQL file specifying the table's schema
\item CSV file to seed the table's values
}
}
\note{
It is assumed that the \code{tbl_name} mirrors both the basename of the CSV file
and the SQL file (excluding extensions).
}
