% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{data-utils}
\alias{data-utils}
\alias{pull_unique}
\title{Data Manipulation Utilities}
\usage{
pull_unique(df, var)
}
\arguments{
\item{df}{a provided \code{\link[=data.frame]{data.frame()}}}

\item{var}{character/numeric - quoted named of a variable from \code{df} or its
numerical index.}
}
\value{
\itemize{
\item \code{pull_unique} returns a character vector of unique, sorted values from specified column
}
}
\description{
Data Manipulation Utilities

Pull Unique Values from a dataframe
}
\examples{
df <- data.frame(let = rep(letters, 2), num = rep(c(1:26), 2))
pull_unique(df, 1)
pull_unique(df, "num")
}
\concept{Data Utilities}
\keyword{cleanse}
\keyword{data,}
\keyword{munge,}
\keyword{utility,}
