% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dates.R
\name{dates}
\alias{dates}
\alias{end_of_month}
\alias{start_of_month}
\alias{extract_date}
\alias{elapsed_months}
\title{Date Utility Functions}
\usage{
end_of_month(date)

start_of_month(date)

extract_date(string)

elapsed_months(end_date, start_date)
}
\arguments{
\item{date}{character or date representation of a date.}

\item{string}{string to extract a date from}

\item{end_date}{end date}

\item{start_date}{start date}
}
\value{
\itemize{
\item \code{end_of_month} returns last day of the month given.
\item \code{start_of_month} returns the first day of the month given.
\item \code{extract_date} returns a date object extracted from the provided string.
}

numeric
}
\description{
A set of helper functions for dealing with dates in a typical actuarial
analysis context.

Derive the number of months elapsed between two dates.
}
\examples{
# character input
start_of_month("2020-08-13")
end_of_month("2017-10-20")

# can handle human-readable dates also
start_of_month("July 7, 1999")
end_of_month("February 5, 2019")

# date input
start_of_month(as.Date("2020-08-13"))
end_of_month(as.Date("2020-10-20"))
}
\seealso{
\code{\link[=as.Date]{as.Date()}}, \code{\link[lubridate:round_date]{lubridate::ceiling_date()}}, \code{\link[lubridate:round_date]{lubridate::floor_date()}},
\code{\link[flipTime:AsDate]{flipTime::AsDate()}}
}
\concept{Date Utilities}
\keyword{dates}
