% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-logit.R
\name{dnorm_logit}
\alias{dnorm_logit}
\title{dnorm_logit}
\usage{
dnorm_logit(p, mean_logit, sd_logit)
}
\arguments{
\item{p}{probability at which to evaluate the density function}

\item{mean_logit}{Mean of logarithmic values}

\item{sd_logit}{Standard Deviation of logarithmic values.}
}
\value{
z-value representing output of a given value of X
}
\description{
Lognormal Density Function
}
\examples{
dnorm_logit()
}
