% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{doc_data}
\alias{doc_data}
\title{Document Datasets}
\usage{
doc_data(
  obj,
  title = deparse(substitute(obj)),
  description = "DATASET_DESCRIPTION",
  write_to_file = TRUE,
  ...
)
}
\arguments{
\item{obj}{object to document}

\item{title}{Title}

\item{description}{Description}

\item{write_to_file}{Logical}

\item{...}{N/A}
}
\value{
silently returns the doc_string
}
\description{
Creates skeleton to document datasets via \code{roxygen2}.
}
\examples{
library(lossrx)
data(losses)
string <- doc_data(losses, "Loss Data", "Claims Data", FALSE)
cat(string)
}
