% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{elapsed_months}
\alias{elapsed_months}
\title{Elapsed Months}
\usage{
elapsed_months(end_date, start_date)
}
\arguments{
\item{end_date}{end date}

\item{start_date}{start date}
}
\value{
numeric
}
\description{
Derive the number of months elapsed between two dates.
}
