% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-data.R
\docType{data}
\name{exposures}
\alias{exposures}
\title{exposures}
\format{
A \code{data.frame} with 855 rows and 5 variables:
\describe{
\item{\code{member}}{character. DESCRIPTION.}
\item{\code{program_year}}{double. DESCRIPTION.}
\item{\code{department}}{character. DESCRIPTION.}
\item{\code{payroll}}{double. DESCRIPTION.}
\item{\code{miles}}{double. DESCRIPTION.}
}
}
\usage{
exposures
}
\description{
Exposure data.
}
\keyword{datasets}
