% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_date}
\alias{extract_date}
\title{Extract Date}
\usage{
extract_date(string)
}
\arguments{
\item{string}{character string}
}
\value{
date object
}
\description{
Extracts dates from a character string with format "mm-dd-yyyy" or "mm/dd/yyyy".
}
\examples{
extract_date("data-raw/lossrun-as-of-2019-01-31.xlsx")
}
