% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-strings.R
\name{extract_num}
\alias{extract_num}
\title{Extract numbers from a string}
\usage{
extract_num(string)
}
\arguments{
\item{string}{String to pull numbers from}
}
\value{
String of numbers
}
\description{
Extract numbers from a string
}
