% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-feedback.R
\name{indent}
\alias{indent}
\title{Indent}
\usage{
indent(x, first = "  ", indent = first)
}
\arguments{
\item{x}{The string passed to various \code{msg_} functions.}

\item{first}{what to indent with - defaults to \code{"  "}.}

\item{indent}{indentation of next line - defaults to \code{first}}
}
\value{
string
}
\description{
Indentation around various \code{msg_} feedback functions.
}
\seealso{
Other Feedback Utilities: 
\code{\link{feedback}},
\code{\link{inform}()}
}
\concept{Feedback Utilities}
