% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-logit.R
\name{logit}
\alias{logit}
\title{logit}
\usage{
logit(p)
}
\arguments{
\item{p}{probability, where 0 < \emph{p} < 1.}
}
\value{
Log of p/(1-p)
}
\description{
The \code{logit} function in mathematics is the \code{quantile} associated with the standard
\href{https://en.wikipedia.org/wiki/Logistic_distribution}{logistic distribution}.
}
\details{
The \code{logit} is used widely as a method of transforming data for modelling purposes.
}
\examples{
logit(.25)
}
