% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actuary-loss_run.R
\name{loss_run}
\alias{loss_run}
\title{loss_run}
\usage{
loss_run(val_date, trans_dat)
}
\arguments{
\item{val_date}{date the valuation date of the loss run.  Claim values from \code{trans}
will be values as of the \code{val_date}}

\item{trans_dat}{data frame of claims transactions}
}
\value{
data frame of claims (1 claim per row) valued as of the \code{val_date}
}
\description{
view losses as of a specific date
}
