% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toproper}
\alias{toproper}
\title{To Proper}
\usage{
toproper(
  string,
  replace_underscores = TRUE,
  underscore_replacement = " ",
  return_as = c("titlecase", "uppercase", "lowercase", "asis"),
  uppers = c("Tpa")
)
}
\arguments{
\item{string}{string to manipulate on}

\item{replace_underscores}{Logical: if \code{TRUE} replaces all underscores
with specified \code{underscore_replacement} argument's value.}

\item{underscore_replacement}{Character: if argument \code{replace_underscores}
equals \code{TRUE}, will replace all "_"'s with specified string.}

\item{return_as}{How should the string be returned? Options are:
\itemize{
\item{"titlecase"}: Applies \code{stringr::str_to_title}.
\item{"uppercase"}: Applies \code{toupper}.
\item{"lowercase"}: Applied \code{tolower}.
\item{"asis"}: No manipulation. Returns as is.
}}
}
\value{
"Proper" string
}
\description{
To Proper
}
\examples{
s <- "variable_a is awesome"
toproper(s)

}
\seealso{
\code{\link[stringr]{str_replace}}.
}
