% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangles.R
\name{tri}
\alias{tri}
\title{\code{tri} - Actuarial Triangle}
\usage{
tri(origin, age, value)
}
\arguments{
\item{origin}{Numeric/Character - Represents the 'origin' or period who's
beginning is the start of a period (i.e. Accident Year).}

\item{age}{Numeric - Age of Development in Months}

\item{value}{Numeric - Value representing Paid Dollars,
Incurred/Reported Dollars, or Claim Counts}
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble()}} structure with class "tri"
}
\description{
Loss Development Triangle Structure
}
\examples{
tri(2020, 12, 50000)
}
