% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actuary-claim_history.R
\name{view_claim_history}
\alias{view_claim_history}
\title{View an Individual Claim's History}
\usage{
view_claim_history(claim_id, claims_data = claims_transactional)
}
\arguments{
\item{claim_id}{Claim ID}

\item{claims_data}{Dataset}
}
\value{
a list containing a) claim details b) transactional history c) interactive timeline
}
\description{
This is a comprehensive function that allows the user to gain valuable insights
about an individual claim's historical transactions.
}
