% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-feedback.R
\name{inform}
\alias{inform}
\title{Inform}
\usage{
inform(...)
}
\arguments{
\item{...}{Passed to \code{\link[rlang:abort]{rlang::inform()}}}
}
\value{
feedback in console
}
\description{
A wrapper around \code{\link[rlang:abort]{rlang::inform()}} for providing feedback to developers using
this packages functions.
}
\seealso{
\code{\link[rlang:abort]{rlang::inform()}}

Other Feedback Utilities: 
\code{\link{feedback}},
\code{\link{indent}()}
}
\concept{Feedback Utilities}
