% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actuary-interp.R
\name{interp}
\alias{interp}
\alias{interp.dblexp}
\alias{interp.exp}
\alias{interp.linear}
\title{\code{interp} - Interpolate}
\usage{
interp(new_age, cdf_array, age_array, cutoff = 450, method = 3)

interp.dblexp(new_age, age_high, age_low, cdf_high, cdf_low)

interp.exp(new_age, age_high, age_low, cdf_high, cdf_low)

interp.linear(new_age, age_high, age_low, cdf_high, cdf_low)
}
\arguments{
\item{new_age}{integer - value of the new age whose CDF is to be interpolated}

\item{cdf_array}{numeric vector of CDFs (usually representative of the selected factors)}

\item{age_array}{numeric vector of ages corresponding to the supplied \code{cdf_array}}

\item{cutoff}{the largest possible age, after which, no interpolation is performed}

\item{method}{integer - must be 1, 2, or 3 where 1 represents linear, 2 represents exponential,
and 3 represents double exponential. Defaults to 3, but falls back onto 1 if necessary.}

\item{age_low, age_high}{Low and High ages}

\item{cdf_low, cdf_high}{Low and High CDFs}
}
\value{
derived numeric value for the supplied \code{new_age}'s CDF
}
\description{
Interpolate Cumulative Loss Development Factors
}
\details{
Actuaries often have to \code{interpolate} values in between the selected LDFs/CDFs
in order to derive development factors at a variety of possible ages of maturity.

This generic function comes with three possible \code{method}s:
\enumerate{
\item Linear Interpolation
\item Exponential Interpolation
\item Double Exponential Interpolation
}
}
\section{Functions}{
\itemize{
\item \code{interp.dblexp}: Double Exponential Interpolation

\item \code{interp.exp}: Exponential Interpolation

\item \code{interp.linear}: Linear Interpolation
}}

\examples{
cdfs <- c(3.579, 2.866, 2.489, 2.121, 1.876, 1.543, 1.222, 1.150, 1.109, 1.005, 1.0025)
ages <- seq(from = 12, to = (length(cdfs) * 12), by = 12)

interp(14, cdfs, ages)
interp(12, cdfs, ages) == cdfs[[1]]
interp(27, cdfs, ages, method = 2)
}
