% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-lossrx.R
\docType{package}
\name{lossrx-package}
\alias{lossrx}
\alias{lossrx-package}
\title{lossrx: Actuarial Loss Development and Reserving with R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Actuarial Loss Development and Reserving Helper Functions and ShinyApp.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig/lossrx}
  \item \url{https://docs.jimbrig.com/lossrx}
  \item Report bugs at \url{https://github.com/jimbrig/lossrx/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

}
\keyword{internal}
